+++ app/src/main/java/com/github/gericass/world_heritage_client/App.kt
@@ -2,6 +2,8 @@ package com.github.gericass.world_heritage_client
 
 import android.app.Application
 import com.facebook.stetho.Stetho
+import com.github.gericass.world_heritage_client.di.Modules.apiModule
+import com.github.gericass.world_heritage_client.di.Modules.repositoryModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 import timber.log.Timber
@@ -14,6 +16,7 @@ class App : Application() {
         startKoin {
             // Android context
             androidContext(this@App)
+            modules(listOf(apiModule, repositoryModule))
         }
         if (BuildConfig.DEBUG) {
             Stetho.initializeWithDefaults(this)
