+++ data/src/main/java/com/github/gericass/world_heritage_client/data/AvgleRepository.kt
@@ -1,0 +1,13 @@
+package com.github.gericass.world_heritage_client.data
+
+import com.github.gericass.world_heritage_client.data.model.Categories
+import com.github.gericass.world_heritage_client.data.model.Videos
+
+interface AvgleRepository {
+
+    suspend fun getCategories(): Categories
+
+    suspend fun getVideoByCategory(page: Int, categoryId: String): Videos
+
+    suspend fun searchVideo(page: Int, keyword: String): Videos
+}
